+++ lib/fastlane/plugin/copy/helper/copy_helper.rb
@@ -1,9 +1,0 @@
-module Fastlane
-  module Helper
-    class CopyHelper
-      def self.copy(source, destination)
-        Actions.sh("cp", source, destination)
-      end
-    end
-  end
-end