+++ lib/fastlane/plugin/file_manager/actions/copy_files_action.rb
@@ -1,15 +1,14 @@
 require "fastlane/action"
-require_relative "../helper/copy_helper"
 
 module Fastlane
   module Actions
-    class CopyAction < Action
+    class CopyFilesAction < Action
       def self.run(params)
-        Helper::CopyHelper.copy(params[:source], params[:destination])
+        Actions.sh("cp", params[:source], params[:destination])
       end
 
       def self.description
-        "Copy file"
+        "Copy files"
       end
 
       def self.authors