+++ spec/copy_files_action_spec.rb
@@ -1,9 +1,9 @@
-describe Fastlane::Actions::CopyAction do
+describe Fastlane::Actions::CopyFilesAction do
   describe "#run" do
     it "executes cp command with source and destination" do
       expect(Fastlane::Actions).to receive(:sh).with("cp", "test.rb", "lib")
 
-      Fastlane::Actions::CopyAction.run(source: "test.rb", destination: "lib")
+      Fastlane::Actions::CopyFilesAction.run(source: "test.rb", destination: "lib")
     end
   end
 end