+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.1.0 - 2019-05-02
+## [0.2.0] - 2019-05-04
 ### Added
-- copy action
+- `remove_files` action
+
+### Changed
+- Rename gem to `fastlane-plugin-file_manager`
+- Rename from `copy` to `copy_files`
+
+## [0.1.0] - 2019-05-02
+### Added
+- `copy` action
+
+[0.2.0]: https://github.com/jasonnam/fastlane-plugin-file_manager/compare/0.1.0...0.2.0
+[0.1.0]: https://github.com/jasonnam/fastlane-plugin-file_manager/releases/tag/0.1.0